MOve tikv-store client to its own module
